[Paropt] Distinguish map-chains with/without explicit map-types.
This patch theoretically improves stability as a follow-up to
https://git-amr-2.devtools.intel.com/gerrit/#/c/313416/ but we
don't have any tests that actually fail without it (yet?).

It is helpful in distinguishing map-chains with "0" map-type sent
by the frontend, and those with the old style without any explicit
map-type (which used AGGR/AGGRHEAD or array section representations).
The default value of map-type is "0" for old-style chains as well.

Once support for old style map-chains is removed, we won't need this.

Change-Id: Ia962bda8363c8eb1285d9a167ebd58858cd4c000
JIRA: CMPLRLLVM-26522, CMPLRPTA-683
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/313419
Reviewed-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>