[AddSubReassociation] Fix crash caused by assertion that all tress in same cluster extended to same degree through shared leaves.
The assertion happened in the part of the pass algorithm that tries to extend
trees within same cluster via turning their shared leaves into tree nodes.
The algorithm asserts that all the trees in the cluster extended at the same
degree. The expectation is reasonable but the problem is that there is a tree
size limit parameter and when trees are not all of same size the bigger one hits
the limit first and thus extends less than others. The problem fixed via
(1) performing tree size check for all trees in cluster to find out the
remaining room to grow before the biggest among the trees reaches its maximum
allowed size and (2) check against that growth limit instead of maximum tree size
when growing a tree so that all the trees would be limited exactly same way
regardless of their initial size.

Fixes CMPLRLLVM-26541.

Change-Id: I137ceb57aa4a4efd5c0b2ecf5c8ac7c81d22b5fa
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/313661
Reviewed-by: Pushkar Ratnalikar <pushkar.v.ratnalikar@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>