[NFC] [DTrans] Reduce build dependencies among DTrans files
The DTransCommon.h file depended upon every DTrans pass, and every DTrans pass
and several non-DTrans components depended on DTransCommon.h, causing any
change to a DTrans analysis or transformation header to trigger a rebuild
of a lot of unnecessary files. This change restructures the list of includes
needed for the pass builders to be separate from the other definitions to
break this cyclic dependency, so that changing a transformation pass header
does not cause everything to rebuild.

Doing this revealed several locations where passes were needing headers
that were not directly included.

Change-Id: Ibce3774e62eeaa90a1d1629e79f34cfe728a7675
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/313908
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>