[DTRANS] CMPLRLLVM-26590: Enable DTrans resolve types when the types are used with llvm.intel.subscript intrinsic
This change set enables DTrans resolve types when the types are used with the Intel
subscript intrinsic. Currently, resolve types won't work for a type if it use
as parameter for a function declaration. The compiler considers that the
llvm.intel.subscript intrinsic is another function declaration, but actually it
is lowered later as a series of GEPs. We know that the intrinsic won't change
the type, therefore we can apply resolve types for it.

Change-Id: I820a434424c55ee886196caee4f051a79b0d333b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/313887
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>