[DTRANS] CMPLRLLVM-26590: Remove redundant BitCast instructions generated by resolve types
There is a chance that merging two types can generate a redundant BitCast instruction. This
is a BitCast where the source and the destination are the same type. This change set adds
a post-processing function for DTrans resolve types. The new function check for redundant
BitCast instructions, replace the users with the pointer operand and delete them.

Change-Id: I1d93ff68b5a0d18ce8f7773e4a987c16a93513c4
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/314207
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Sergey Kanaev <sergey.kanaev@intel.com>