CMPLRLLVM-26638: Disable devirtualization from exported symbols
The community introduced a change set that disables the devirtualization
for exported symbols when setting the visibility (https://reviews.llvm.org/D91583).
The difference is that the community depends on an assumption flag to decide
on how the visibility will be handled. We replaced the assumption flag with
the result from the whole program analysis.

This change set enables the collection of the symbols' resolutions from the linker
and stores them in the whole program utility. Then we collect which symbols were
marked as export and disable the devirtualization for them.

Change-Id: I889ac6adc3e11dad92dce013c92826346e67189b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/314277
Reviewed-by: Masten, Matt <matt.masten@intel.com>
Tested-by: Masten, Matt <matt.masten@intel.com>