Implement attribute allow_cpu_features
Similar to how may_i_use_cpu_features_ext works, this attribute uses a
pair of uint64_t parameters to identify the features.  However, in this
case, it actually allows the use of the feature in the function (similar
to how attribute 'target' does).

The motivation for this feature is to allow customers to specify a list
of permitted features in a function in a way that would permit use in a
template, as requested in CMPLRLLVM-26635.

Change-Id: I8229a07b63356870bfe2d256d205f2d4470acd3d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/314540
Reviewed-by: Keane, Erich <erich.keane@intel.com>
Tested-by: Keane, Erich <erich.keane@intel.com>