[DTrans - OpaquePtr] Update specification and reading of DTrans metadata
This updates the description of the proposed DTrans metadata format, and
the metadata parsing functionality to preserve and read what is being worked
on with front-end team.

This change reads either the updated or old metadata format. Subsequent
change sets will update the LIT tests to be in the new format and remove
reading of the prior experimental format.

Change-Id: I361cb646ca78fe72477e73a58a2e81a3bdb4c3bf
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/314390
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>