[DTrans - OpaquePtr] Update TypeMetadataReaderTest pass to check whether metadata matches expected type
As long as opaque pointers are not being used yet, the types reconstructed from
metadata annotations can be checked against the expected types when checking
whether the information needed for DTrans is present. This change set adds
checks of this sort when running 'opt' with the 'dtrans-typemetadatareader' test
pass. This can help identify locations where metadata is lost or mistranslated
before it gets to DTrans.

Change-Id: Ie45ecc8a82acea19f3ef793028df1fe08f4d1294
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/314391
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>