CMPLRLLVM-26696: Fix for whole program assert in the release compiler
The whole program assert wasn't working in the release compiler
because an early return was triggered if whole program wasn't achieved.
This change set makes sure that the whole program assert happens
after analyzing the function calls and the aliases. It enables the
assertion in the release compiler.

A test case for this is in Analysis/Intel_WP/whole_program_new_pass_manager.ll.
The test case was failing in the release compiler but it is working now.

Change-Id: I5c1277223ffa6cacc2d3cea1a6b2151a11c27955
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/314609
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>