[AndersensAA] CMPLRLLVM-26345: AndersAA tries to use operand of bitcast when AndersAA doesn't know about the bitcast.
AndersAA doesn’t know about a BitCast If the BitCast instruction is created
after AndersAA results are computed. AndersAA tries to use the operand
of the BitCast instruction if it know about it.
This change is needed to fix a performance regression for 544.nab due to
pulldown. More details of the issue is added to the CMPLRLLVM-26345.

Change-Id: I73bd0014e1f8b583563bff45b7529c5966d8e104
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/314738
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>