Remove registry for _alloca
This was used to solve a cl-link issue. Now we have already switch the
linker to lld-link, so we don't have to manually register _alloca.

And icx identifies _alloca as a builtin (which is expected), but doesn't
allow its function pointer passing around (only direct call is allowed).

Therefore we just remove it.

Change-Id: Id4a27efd1a939c317083212160a8369eb643448a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/314817
Tested-by: Yilong Guo <yilong.guo@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>