[SLP] Fix crash when matching associative reduction for integer min/max.
Associative reduction matcher in SLP begins with select instruction but when
it reached call to llvm.umax (or alike) via def-use chain the latter also matched
as UMax kind. The routine's later code assumes matched instruction to be a select
and thus it merely died on the first encountered cast that did not fit.

Differential Revision: https://reviews.llvm.org/D98432

Change-Id: I51b81a5f8bff5ef2986339ef7a13d402b0e1fd36
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/315835
Reviewed-by: Dmitriev, Valery N <valery.n.dmitriev@intel.com>
Reviewed-by: Masten, Matt <matt.masten@intel.com>
Tested-by: Masten, Matt <matt.masten@intel.com>