[DTrans] Add 'experimental' intrinsics to the list of innocuous intrinsics
Previously, llvm.experimental.* intrinsics were not considered for inclusion
in the table of safe intrinsics. However, the compiler is now using a
llvm.experimental.noalias.scope.decl call on 523/623. This change adds the set
of llvm.experimental, except for the vector ones to the list of safe intrinsics.

This addresses a performance regression reported in: CMPLRPTA-713 & CMPLRPTA-719

Change-Id: Ia01c144631b325089e14bdd914651cfad745055f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/315174
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>