[NFC][VPlan] Add getMemoryRef interface in VPLoadStoreInst
The changes in this patch add an interface to get the HIR memory reference
for VPLoadStoreInst. The interface is essentially extracted from what was
being used in getUnderlyingNonDbgMetadata so that the same interface can
be used in subsequent patches to extend VLS analysis for HIR instructions
with multiple memory references.

Change-Id: Ic5b97af86b3a6a52dedb979578f23c005eb914bb
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/315539
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: Guggilla, Satish <satish.guggilla@intel.com>