[Paropt] Handle const-expr section ptrs which are used inside the region.
Previously, renaming/replacement of section-pointer operands was happening
for the full region, so if the same constant-expr was used in the region,
that would be replaced with the renamed section-pointer operand, instead
of getting computed using the base-pointer operand.

Now, renaming of section-pointer operands only happens on the begin
directive of the region.

Change-Id: Icd6556899b56d89440cba22014c589cfe0532f11
JIRA: CMPLRLLVM-26811
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/315659
Reviewed-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>