CMPLRLLVM-26865: Fix for DTrans analysis to check when a Sub instruction is used for allocating a memory space
After the pulldown, the IR generated uses a PtrToInt instruction to calculate the
difference between two pointers rather than a BitCast and Load. This change set
extends the DTrans analysis to check if a new entry in an STD vector is added
and the size was computed with PtrToInt instructions.

Change-Id: Ibc2ab3e1f4dff371176322deed05a1965ec3fa96
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/315852
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>