[X86] Fix compile time regression of D93594.
D93594 depend on the dominate tree and loop information. It increased
the compile time when build with -O0. However this is just to amend the
dominate tree and loop information, so that it is unnecessary to
re-analyze them again. Given the dominate tree of loop information are
absent in this pass, we can avoid amending them.

Differential Revision: https://reviews.llvm.org/D98773

Change-Id: I36af200daf643c419cd4a0f32f7445c5d970bb19
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/317531
Reviewed-by: Luo, Yuanke <yuanke.luo@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>