Support replacing aligned vector moves with unaligned moves when avx is enabled.
With AVX the performance for aligned vector move and unaligned vector move on X86
are the same if the address is aligned. In this case we prefer to use unaligned
move because it can avoid some exceptions.
"-fuse-unaligned-vector-move" and "-fno-use-unaligned-vector-move" are added to
enable this preference. We will enable this when "-fintel-compatibility" is enabled.
Which means "-fuse-unaligned-vector-move" will as a default option in ICX.

Doing the transform in assembly because we only want to changing the output of
instructions without any change of IR and flag. This is also how ICC implements.

Change-Id: I49eb736a1994d39bb332a0826a428fa694fd50b8
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/316123
Reviewed-by: Toguchi, Michael D <michael.d.toguchi@intel.com>
Reviewed-by: Luo, Yuanke <yuanke.luo@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Sergey Kanaev <sergey.kanaev@intel.com>