Add new command-line switch to include the filename (not the path) by default with __builtin_FILE
Clang currently shows the file name as specified on the command-line into the
final assembly when using __builtin_FILE, which is used for source-location
purposes. CMake always uses full-paths, so our user's programs are having
their full build-machine paths displayed.

This is a security problem.

This patch adds a new command-line switch to control this behavior that
__builtin_FILE should only include the filename (not the path) by default
and keeps the switch 'on' by default in -fintel-compatibility mode.

This fixes https://jira.devtools.intel.com/browse/CMPLRLLVM-26876.

Change-Id: I37ece9c38dd4983f7fdd55cc81f1f06a2215919d
Signed-off-by: Soumi Manna <soumi.manna@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/316196
Reviewed-by: Keane, Erich <erich.keane@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>