[DTRANS] CMPLRLLVM-26865: Add the safety bit FieldAddressTakeForCallSite
This is the second part for fixing the DTrans analysis for arrays with
constant entries. This change set replaces the DTrans' safety mask
FieldAddressTaken with three new definitions: FieldAddressTakenMemory,
FieldAddressTakenCall and FieldAddressTakenReturn. FieldAddressTakenMemory
means that the address of field in a structure is written.
FieldAddressTakenCall is used to identify when the address of a field is
passed to a callsite, and FieldAddressTakenReturn means that the address
of a field is returned by a function.

The motivation for this change set is to have a better explanation on
the reason for marking 'FieldAddressTaken' to the information of a
structure (StructInfo in DTrans). This was noticed because the function
std::vector::push_back was inlined when the libraries of GCC 7.x are used.
When the default GCC was updated to GCC 10.x then this function is not
inlined anymore because the inline cost is higher since the implementation
changed. The field of a structure can be added to a std vector through
push_back and produce 'Field Address Taken' in DTrans.

Change-Id: I2738fedf87303d53c2a32b61b8cc1b1c91936ff2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/316251
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Sergey Kanaev <sergey.kanaev@intel.com>