CMPLRLIBS-3774: prepare for instrumenting SPIR-V device RTL with ITT.
Some changes are still only in intel/llvm.

We can already start linking ITT libraries by default,
so I changed the driver to do that. I also changed the linking
for device RTL, because it has to follow the same rules as
the other libdevice parts. We will have to enable the instrumentation
pass in the driver, when github changes merge into xmain.

The paropt change is to make sure that the work_group_barrier call
is recognized by the instrumentation pass (not in xmain yet).

The device RTL changes are currently disabled, but I verified
in a local ws that the instrumentation pass works.

Change-Id: I681d2052aae597d1a98bf68bdb66cf07815b4494
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/316442
Reviewed-by: Toguchi, Michael D <michael.d.toguchi@intel.com>
Tested-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>