[VPlan] Correction for emitting VPVectorTripCount VPInstruction.
The error is in that the VPVectorTripCount instruction is emitted
at the beginning of the loop preheader always, assuming that its
operand (loop upper bound) is always calculated before that preheader.
That causes the error when its operand (loop upper bound) is calculated
in the loop preheader or in the block that dominates it and the parent
of the loop upper bound is in the same VPlan.
The correction is to check the placement of the loop upper bound and
emit VPVectorTripCount before it.

JIRA is CMPLRLLVM-26959.

Change-Id: I25adad1938611b13a19db0695a6158ae9614c783
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/316638
Reviewed-by: Artem Gindinson <artem.gindinson@intel.com>
Tested-by: Artem Gindinson <artem.gindinson@intel.com>