[VPlan] Introducing VPPushVF and VPPopVF VPInstructions.
The new VPInstructions are placeholders to change vector and unroll factors in
common merged CFG. They are inserted at the beginnning and in the end of each
VPlan during CFG merging. Encountering the VPPushVF CG saves the current VF/UF on
the dedicated stack and sets new their values. The VPPopVF is the reverse operation,
the values from the top of stack replace the current VF/UF and the top of stack is
discarded.

Change-Id: Ic60fa985751307c3224ed6a8b15593354a75faa3
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/317103
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>