[VPlan] Setup call argument types correctly for getting TTI based cost in VPlan cost modeling
VPlan cost modeling was not using transform function correctly which ended
up passing an empty argument types vector to TTI based cost modeling interfaces
resulting in a compiler crash. The use of transform is unnecessary and this has
been replaced with a loop that sets up the argument types vector.

JIRA :: CMPLRLLVM-27036

Change-Id: I681e1502f66144f342c7272147c1f8a9220d9952
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/317411
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: Guggilla, Satish <satish.guggilla@intel.com>