[Paropt] Add support for UDR with initializers.
With the change, Paropt replaces the uses of
kmpc_task_reduction_init API with kmpc_taskred_init, which uses
a two argument initializer callback function, instead of a
single argument one. This allows support for calls to UDR
initializer functions which need the original operand to be
passed in.

Change-Id: I3d140c0c4b28ab2e15220b1f94f02b50e10f2f43
JIRA: CMPLRLLVM-26976
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/317496
Reviewed-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>