Move OCL Customization before community modifies type variable
This fixes CMPLRLLVM-27088.  It appears that the community code depends
on knowing that the variable 'ty' is not used after the 'else' block, so
they re-use it and 'de-pipe' it at one piont.  This patch simply moves
the code we have that uses it above the modification so that we aren't
affected by that change.

Change-Id: Ic352797615444820891893f3bfd5304bf65c7b04
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/317746
Reviewed-by: Artem Gindinson <artem.gindinson@intel.com>
Tested-by: Artem Gindinson <artem.gindinson@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/317765
Reviewed-by: Keane, Erich <erich.keane@intel.com>