CMPLRLLVM-26865: Disable the simplifying load into bitcast if it can generate an upcasting
This change set disables converting a load instruction into a bitcast if it
can produce an upcasting, DTrans is enabled and loopopt passes hasn't run.
The simplification will be turned on once the pass manager has processed
the loopopt transformations in LTO. The motivation for disabling the
simplification is that generating a bitcast that could be an upcasting
can complicate the DTrans analysis. It can could produce misleading results.

Change-Id: I8191616906551003550727839f2a32b956399a15
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/317870
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Lin Jr, Chang-Sun <chang-sun.lin.jr@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>