[VPlan][HIR] Fix revectorization of address-of DDRef for gather/scatter memrefs.
Currently VPlan HIR vectorizer codegen generates simple widened
address-of DDRef while creating a memref that corresponds to
gather/scatter access. However this would lead to invalid vector
GEP in LLVM-IR if we are dealing with VectorType in incoming HIR,
as GEP's return value VectorType size would not match that of the
indices' VectorType. We need to flatten and replicate the base
pointer of widened address-of DDRef in-order to prevent generation
of such invalid GEPs during HIR-CG. This patch adds support to
handle such revectorization scenarios in HIR vector codegen.

JIRA : CMPLRLLVM-27000, CMPLRLLVM-27069

Change-Id: I35f3c7d7b2cf2a6dcc59de16501f57cce8f7d9a6
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/317811
Tested-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>