Fix prod build with asserts disabled
LLVM has a pre-processor macro LLVM_ENABLE_DUMP that allows some
debug-like code to be compiled even when NDEBUG is defined. This macro
is enabled by default for prod builds, but we had some Intel-specific
code that was incorrectly mixing the "#if !defined(NDEBUG) ||
defined(LLVM_ENABLE_DUMP)" with "#ifndef NDEBUG" causing compilation
errors if a prod build is attempted with assertions disabled.

This patch addresses CMPLRLLVM-22922.

Change-Id: I1e4b2f018cfbe9d83690d6d7df971157a1d49e74
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/317810
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>