[DTRANS] CMPLRLLVM-26963: Add the DTInfo for base types
This change set adds the creation of the type-info for structures with
base name. For example, consider that the following structures are in
the module:

  %class.simple = type { i32, i64 }
  %class.simple.1 = type { i32, i64 }

Assume that we are generating the type-info for %class.simple.1. The
base name will be %class.simple. In this case we need to generate the
type-info for %class.simple too. The issue is that there is a chance
that the base-named structure is never used in the module, which means
that DTrans analysis will never generate a type-info for it when
traversing the module. DTrans-opt-base collects the base-named structure
and since the type-info wasn't generated then any transformation that
tries to access it will produce a segmentation fault.

Change-Id: I95ba184ee067dd15e77c2279c3ac47005bc51893
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/318022
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>