Using the loop with MinRef as the extracted loop in HIR Store Result Into Temp Array
Originally, when we do bulk loop carried scalar replacement, we use the first loop
to create the extracted loop because we suppose the minref is in the expression tree
of the first loop. In 503.bwaves, the minref is in the third loop. This changeset finds
the loop with minref and use the corresponding expression tree to create the extracted
loop.

Change-Id: I9a9ba571d522c3a36057100400578dede0b06ae3
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/318290
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>