[VPlan] Correction to emit induction close form calculation.
When induction for which we generate close form is used in the
latch condition and the latch condition is used more than once,
we need to duplicate the latch condition before replacing uses
of induction with the new one calculated by close form. The old
condition and all its uses should be kept except one in the
latch, because we change induction and thus we change that latch
condition.

JIRA: CMPLRLLVM-27146.

Change-Id: I2fd8c0209518f002c0b0447be6fb4430a24075d1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/318462
Tested-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Dmitriev, Valery N <valery.n.dmitriev@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>