[VPlan][CallVecDecisions] Prepare CallVecDecisions to run for merged VPlan CFG.
Currently CallVecDecisions is designed to be a VF-dependent
analysis. Consumers of the analysis provide a specific VF
which is used to analyze all VPCalls present in a given VPlan.
However this design is not futureproofed for upcoming VPlan
features where different VPlans can be fused to form a single
merged CFG (performed by VPlan CFGMerger transform). In such
merged CFGs different VPCalls need to be analyzed with
different VFs tracked by VPPush/PopVF instructions. This patch
updates CallVecDecisions to be flexibly used in 2 modes -

1) runForVF - VF is explicitly provided by caller
2) runForMergedCFG - VF for each call is determined from CFG

Change-Id: Ibc7dcf0ba1ab12017e43b76306e09f2da4cf42c2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/318866
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>