CMPLRLLVM-27232: Argument promotion should not cause new/delete of callsites in the inlining report.
This introduces some new methods to the InlineReportCallSite class:
  replaceCallBaseWithCallBase
  cloneCallBaseToCallBase
that allow us to record that new callsite is a replacement of or is a clone of
an existing callsite. Using these methods will remove the pairs of deleted and
newly created callsites in the report.

A CallbackMap is now used to track callbacks so they can be added and removed
in constant time.

We also introduce initModule and rename initFunctionForPartialInlining to
initFunctionClosure so that they can be used in a more general context for
module level and SCC level passes.

The new methods are applied to ArgumentPromotion, and will be applied to other
optimizations in the future. So far, support exists only for the classic
inlining report. Support for the metadata inlining report will follow in
future change sets.

Change-Id: I5b08408ed90f0bad51c941dba1915bcc7b322efd
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/318772
Tested-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>