[NFC][VPlan][SVA] Make ScalVec analysis compute cycle to be VF-independent.
VPlan's ScalVec analysis is driven to compute results based on
3 prerequisites - DA, def-use chains, CallVecDecisions. As such
the VF chosen to vectorize a given VPlan doesn't directly influence
SVA results. It is accounted indirectly by CallVecDecisions which
is a VF-dependent analysis. This patch updates SVA's compute
interfaces to be VF-independent thereby allowing clients to invoke
the analysis even on merged VPlan CFG which does not involve a
unique VF.

Change-Id: I3c7446e07b9ac6c72e81cced1c62a2d2bdaa064e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/318867
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>