[NFC][VPlan] Split HIRSpecifics class into two
The first (HIRSpecificsData) is data only, stored inside the VPInstruction.
The second (HIRSpecifics) is a lightweight wrapper to provide access to the
HIRSpecificsData yet maintaining the link to the owning VPInstruction.

That change would allow us to reference the owning VPInstruciton (e.g. for
checking its opcode) from within the methods of the HIRSpecifics class.

Change-Id: I82819795364cf6ef165a29fddaafc69059c8ff8d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/318916
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>