[VPlan][CM] LIT test to demonstrate VLS store group cost being applied twice
Test to demonstrate issue with VLS group cost being applied twice to stores
a group. This happens the first time when we see a new store group. The
second time this happens is when the instruction corresponding to the group
insertion point is seen. Subsequent changes fix this to apply the group cost
only when the instruction corresponding to insertion point is seen.

Change-Id: I8575db491e53923c94e0e2b649f2462705cc29ae
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/318924
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: Guggilla, Satish <satish.guggilla@intel.com>