[Paropt] Add command line flags to allow ignoring of OpenMP regions.
This may be helpful in narrowing-down runfails in programs with
many OpenMP pragmas, if modifying sources to comment out directives
is too compilicated. This could also be used through TC.

The numbering of regions is a bit wonky because every invocation of
buildWRgraph() (even on the same directive) increases the WRegion
node's ID "Number". But the numbers are deterministic, and removal of
one region doesn't change the ID of another.

Three command line flags are available:

-vpo-paropt-ignore-regions-starting-with=N
-vpo-paropt-ignore-regions-up-to=N
-vpo-paropt-ignore-regions-matching=N1,N2,N3...

Where Ns are positive integers representing WRegion IDs.

One example of use is to first use
-vpo-paropt-ignore-regions-starting-with=1, which would cause
Paropt to ignore all WRegions. It would also print the IDs of
all regions ignored (on stderr if using -Rpass=openmp,
or in the opt-report file if using -qopt-report). These IDs
can then be used in subsequent compilations to specifically
disable regions based on their IDs using the three flags listed above.

Change-Id: I64955fb7f4897e82f8ecc0077d6a00921d420da8
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/319133
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>