CMPLRLLVM-27250: Fix bad merge resulting in failing inlining tests
The problem was not a change in inlining heuristics, there was a bad merge
resolution. 'SingleBB' needs to stay defined in the CallAnalyzer, not defined
in the derived class InlineCostCallAnalyzer, as it is used in the CallAnalyzer
in INTEL custom code.

I have verified that all of the LIT tests in Inline and InlineReport
directories pass after this fix.

Change-Id: I635a5b2b214e58a76a9535bafc1ddc8f573018fb
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/319150
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>
Tested-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>