[VPlan][CFGMerge] Implementation of VPInvSCEVWrapper class. Adding support for peel loops in new cfg merge.
This patch implements the VPInvSCEVWrapper class which encapsulates the
the invariant scev. Currently the only use of it is to represent a base-pointer
computation for dynamic-peeling case. This instruction will be used by CG to
expand and get hold of the base-pointer based on which the actual peel-count
would be computed.

The CFG merger is now able to process peel loops inserting VPlanAdapters
for peel loops, merge blocks and the needed checks for peel.

Change-Id: I296eac9173c1e9995344518538e95700eddb065a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/319271
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>