CMPLRLLVM-27287: Add cloning after inlining to the classic inlining report
This adds calls to the inlining report to IPCloning so that cloning after
inlining is reflected in the inlining report. Note that IPCloning can be
run before and after inlining on the IPO link step. Nothing will happen
until the inlining report is first created. The update calls check that
the functions and calls are in the inlining report before attempting to
update them.

I have also removed a line in Intel_InlineReport.cpp that updates the
source language before the report is dumped. This is not needed, as the
source language will not change during compilation.

Change-Id: I6eb78f74243235f7833fe04fa7d42cca5e4ee5f1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/319330
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>