CMPLRLLVM-27321: Add Loop Opt HIR Code Generation to classic inlining report.
As part of its code generation, Loop Opt clones calls and destroys the
original calls. We need to note this to the classic inlining report, so
that the calls do that appear as deleted calls, followed by newly created
calls. In particular, this leads to a loss of the non-inlining reasons.

Change-Id: I84ad34431ccc23c6613f65e5916217ec9f65d4f1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/319441
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Chupina, Olga <olga.chupina@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>