Fix DeclareSYCLBuiltins for round-trip setting.
Community added a round-trip functionality to the command line args so
that we can generate a full set of command line arguments from our
existing state.  At the moment, this is 'assert' build only, but it
seems the intent is to make this perminent for future functionality.

We will have to analyze the full set of our command line args to make
sure we do this right (that is, that it gets set right the 2nd time),
but this one fixes the DeclareSYCLBuiltins to unblock the pulldown's
build.

Change-Id: Ia1dbc752b6c333689fbe28cdf6bcdba90c190954
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/319449
Reviewed-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>
Tested-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>