Changes to perform contraction for a def/use loop pair one base at a time.
The driver now passes all the refs in the def/use loop pair which
need to be contracted to the same base to the contraction utility.

The main advantage of this approach is that the mapping of original ref
to contracted ref can be trivially communicated to the contraction
utility so it becomes a 'pure' function.

The driver is free to perform more complicated mappings which is required
for some arrays in 503.bwaves.

Change-Id: Iddaf8b661e5cdc528ca22922833c1a89abcb6581
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/319757
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>