[NFC] [DTrans] Interface change to have an iterator for CallInfoElementTypes
This is to prepare for being able to use DTransType objects by the
CallInfo classes for supporting the opaque pointer work. This change
makes it so that a consistent interface will be able to be used to
access the CallInfoElementTypes class when changes are made to use
the AbstractType class.

Change-Id: I4055e54f833e01a32130c95f16e682d51d220afc
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/319899
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>