CMPLRLLVM-27446: Update LIT test to reflect community GEP merge changes
The community no longer performs GEP merging for the following sequence:

  %call = call noalias i8* @bar(i64 16)
  %coder = getelementptr inbounds %struct.lzma_next_coder, %struct.lzma_next_coder* %next, i32 0, i32 0
  store i8* %call, i8** %coder, align 8
  %t0 = bitcast i8* %call to %struct.lzma_coder*
  %myint1 = getelementptr inbounds %struct.lzma_coder, %struct.lzma_coder* %t0, i32 0, i32 0
  store i32 15, i32* %myint1, align 8

because removing the bitcast will make it more difficult for phi translation
and MemoryDependencyAnalysis to succeed.

So, I have updated the test case to use @bar instead of @malloc, which is
transformed the same way by InstCombine as it was before the community changes.
See the comment in visitGetElementPtrInst() in
illvm/lib/Transforms/InstCombine/InstructionCombining.cpp for details.

Change-Id: Id801e828da1fb5af200cfc0c47bbdf329d5b11e6
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/320117
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>
Tested-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>